Totally rewritten GUI. A new object is used to manage both the interfaces for Unix and Win (goGUIclass) Fig modification -> Egnos (SBAS) flag added. Tags for panels and missing objects have been set. Some LEDs (presence of file / folders) have been added.
 Brief explanation of the new interface utilization.
-------------------------------------------------------------
In the "gui_goGPS.m" and "gui_goGPS_unix.m" every callback
that modifies some values have to call the  function goGUI.syncFromGUI
E.g.
global goGUI
    goGUI.syncFromGUI(goGUI.idUI.sUName);

In the goGUIclass everything has been changed.
The inizialization begin calling the creator of the object goGUI
than the init function is called:
	- it creates an internal copy of the structure of the interface
	- it imports previous statuses
	- it syncs everithing back to the GUI

The key concept is the id alias.
Inside the function initUIids EVERY object in the GUI is mapped with one id
            i=i+1; id.pSettings     = i;    id2h(i) = obj.goh.uipSettings;
in id2h there is the mapping between the internal id and the handle name

The most important function for any modification are:
	- the event manager:           syncFromGUI
	- the dependencies control:    checkUIdependencies



See also internal documentation
Debugging will follow…