Proposed fix/improvements.
When computing a LS solution epoch by epoch (with double differences) the troposphere error is estimated using the elevation of the satellites and the ellipsoidal elevation of the receiver.
These information are computed from the rough position estimated using a bancroft solution, to improve the estimation of the troposphere error a simple iteration is added to the LS processing.
From a small test I found that 3 iterations seem enough to reach the convergence and optimally estimate the error.

This improvements should be integrated into the trunk in all the functions that have to estimate with LS the tropospheric error.
This tweak could also be applied to the estimation of the ionospheric effect, but I suppose it is less "sensible" to the problem (I didn't make any test about this)