[trunk] Adding new code for integer ambiguity resolution by LAMBDA method, now also for goGPS Kalman filter (code contributed by Andrea Nardo); both LAMBDA2 and LAMBDA3 are supported.
NOTE: LAMBDA3 code cannot be redistributed so it is not included in this commit; it must be downloaded from http://gnss.curtin.edu.au/research/lambda.cfm and put in the folder positioning/lambda/lambda_v3. If LAMBDA3 code is not found, goGPS will issue a warning and switch to LAMBDA2.

The location of all setting files has been moved from ../data/settings to ./settings