Minor update regarding the use of "constellations"
In the case where "constellations" is empty at input, one would run into trouble using other functions that use "constellations" as an input since not all of them have a conditional statement to default to GPS if "constellations" was left empty. I've added "constellations" as an optional output to load_RINEX_nav. This solves this problem without having to modify all the functions that use "constellations".