Bug-fix: LS_DD_code_phase
When the observations of phase are less than the observations of code, the matrix dimension were not right. I need to reduce the observations to use according to the phase observations of the satellites that I have. (keep only satellites that rover and master have in common both in phase and code)