Adding getters methods to Settings for stations and antanna files
Adding getters methods to IO_Settings for:
atx, crd, ocean_loading, met files

When the file are not found in the location indicated in the ini file try to use the ones in the project folder, or in the global folder.

Adding the file of atx, crd, ocean_loading, met relatives to the PPP project in the PPP project folder. They are copied here inside to show the possibility of having stations/antenna files in the folder