🛰 Experimental New feature: backward forward KF
Complementary to the forward - backward solution it's an interesting addition to goGPS.
This feature can only be activated trough the modification of the settings file (no GUI).
Note that as in the FB solution the BF solution is weighting in the proper way only the last epoch... useful in a batch static solution but WARNING: it may introduce wrong weightings.
i.e. the BF is different from the FB solution