Not optimal but better estimation of outliers at the left border
Before this update there were a lot more data flagged as outliers at the left border of the arc. This is due to the fact that I use differentiation in time to detect outliers, I should differentiate in the two directions to discover a possible outliers; with the current approach when the second element of an arc is flagged, also the first is flagged.